refactor: relocate for clarity
